do not overwrite existing cache files

+++ hydra_cached/execution/caching/dumpable.py
@@ -61,7 +61,8 @@ class DumpableInMemoryCache(InMemoryCache[DictConfig, TV]):
     def dump(
         self,
         directory: str = None,
-        constraints: Optional[Iterable[Callable[[DictConfig, TV], Optional[str]]]] = None
+        constraints: Optional[Iterable[Callable[[DictConfig, TV], Optional[str]]]] = None,
+        overwrite: bool = False,
     ):
         if directory is None:
             directory = self.directory
@@ -85,6 +86,12 @@ class DumpableInMemoryCache(InMemoryCache[DictConfig, TV]):
             k_dump = OmegaConf.to_yaml(k) #, sort_keys=True)
             h = joblib.hash(k_dump)
             fn_config = path.join(directory, f'{h}.yaml')
+            # To keep initial file creation timestamps, do not overwrite existing files.
+            if os.path.exists(fn_config) and not overwrite:
+                if self.verbose:
+                    logger.info(f'do not dump cache entry because it already exists: {h} (set overwrite=True to '
+                                f'enforce overwrite)')
+                continue
             with open(fn_config, "w") as f:
                 f.write(k_dump)
             fn_data = path.join(directory, f'{h}.pkl')
